requirements.txt
# Core dependencies (Python 3.12+ recommended for best compatibility)
Pillow>=10.0.0,<13.0.0
opencv-python>=4.8.0,<5.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Type checking and development
mypy>=1.7.0,<2.0.0
black>=23.12.0,<25.0.0
isort>=5.13.0,<6.0.0

# Optional: HTTP server for future API
fastapi>=0.104.0,<0.200.0
uvicorn>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0

